Updated ch05 examples
