+++ 05/01.php
@@ -1,5 +1,5 @@
 <?php
-  echo strrev(" .dlrow olleH"); // Reverse string
-  echo str_repeat("Hip ", 2);   // Repeat string
-  echo strtoupper("hooray!");   // String to upper case
+  echo strrev(' .dlrow olleH'); // Reverse string
+  echo str_repeat('Hip ', 2);   // Repeat string
+  echo strtoupper('hooray!');   // String to uppercase
 ?>
