+++ 05/03.php
@@ -1,6 +1,6 @@
 <?php
-  $names = fix_names("WILLIAM", "henry", "gatES");
-  echo $names[0] . " " . $names[1] . " " . $names[2];
+  $names = fix_names('WILLIAM', 'henry', 'gatES');
+  echo $names[0] . ' ' . $names[1] . ' ' . $names[2];
 
   function fix_names($n1, $n2, $n3)
   {
