+++ 05/04.php
@@ -1,16 +1,16 @@
 <?php
-  $a1 = "WILLIAM";
-  $a2 = "henry";
-  $a3 = "gatES";
+  $a1 = 'WILLIAM';
+  $a2 = 'henry';
+  $a3 = 'gatES';
 
-  echo $a1 . " " . $a2 . " " . $a3 . "<br>";
-  fix_names($a1, $a2, $a3);
-  echo $a1 . " " . $a2 . " " . $a3;
+  function fix_names()
+  {
+    global $a1; $a1 = ucfirst(strtolower($a1));
+    global $a2; $a2 = ucfirst(strtolower($a2));
+    global $a3; $a3 = ucfirst(strtolower($a3));
+  }
 
-  function fix_names(&$n1, &$n2, &$n3)
-  {
-    $n1 = ucfirst(strtolower($n1));
-    $n2 = ucfirst(strtolower($n2));
-    $n3 = ucfirst(strtolower($n3));
-  }
+  echo $a1 . ' ' . $a2 . ' ' . $a3 . '<br>';
+  fix_names();
+  echo $a1 . ' ' . $a2 . ' ' . $a3;
 ?>
