+++ 05/05.php
@@ -1,16 +1,5 @@
 <?php
-  $a1 = "WILLIAM";
-  $a2 = "henry";
-  $a3 = "gatES";
+  include "library.php";
 
-  echo $a1 . " " . $a2 . " " . $a3 . "<br>";
-  fix_names();
-  echo $a1 . " " . $a2 . " " . $a3;
-
-  function fix_names()
-  {
-    global $a1; $a1 = ucfirst(strtolower($a1));
-    global $a2; $a2 = ucfirst(strtolower($a2));
-    global $a3; $a3 = ucfirst(strtolower($a3));
-  }
+  // Your code goes here
 ?>
