+++ 05/06.php
@@ -1,6 +1,5 @@
 <?php
-  include("library.php");
+  include_once "library.php";
 
   // Your code goes here
 ?>
-
