+++ 05/10.php
@@ -1,6 +1,12 @@
 <?php
   $object = new User;
-  print_r($object);
+  print_r($object); echo "<br>";
+
+  $object->name     = "Joe";
+  $object->password = "mypass";
+  print_r($object); echo "<br>";
+
+  $object->save_user();
 
   class User
   {
@@ -8,7 +14,7 @@
 
     function save_user()
     {
-  	  echo "Save User code goes here";
+      echo "Save User code goes here";
     }
   }
 ?>
