+++ 05/13.php
@@ -1,14 +1,12 @@
 <?php
-  $object1 = new User();
-  $object1->name = "Alice";
-  $object2 = clone $object1;
-  $object2->name = "Amy";
-
-  echo "object1 name = " . $object1->name . "<br>";
-  echo "object2 name = " . $object2->name;
-  
   class User
   {
-    public $name;
+    public $name, $password;
+
+    function __construct($name, $password)
+    {
+        $this->name = $name;
+        $this->password = $password;
+    }
   }
 ?>
