+++ 05/14.php
@@ -1,9 +1,9 @@
 <?php
   class User
   {
-      function __construct($param1, $param2)
+    function __destruct()
     {
-      // Constructor statements go here
+      // Destructor code goes here
     }
   }
 ?>
