+++ 05/15.php
@@ -1,9 +1,11 @@
 <?php
-  class Test
+  class User
   {
-    public $name = "Paul Smith"; // Valid
-    public $age = 42; // Valid
-    public $time = time(); // Invalid - calls a function
-    public $score = $level * 2; // Invalid - uses an expression
+    public $name, $password;
+
+    function get_password()
+    {
+      return $this->password;
+    }
   }
 ?>
