+++ 05/18.php
@@ -1,9 +1,17 @@
 <?php
-  class Test
+  Translate::lookup();
+
+  class Translate
   {
-    public $name  = "Paul Smith"; // Valid
-    public $age   = 42;           // Valid
-    public $time  = time();       // Invalid - calls a function
-    public $score = $level * 2;   // Invalid - uses an expression
+    const ENGLISH = 0;
+    const SPANISH = 1;
+    const FRENCH  = 2;
+    const GERMAN  = 3;
+    // ...
+
+    static function lookup()
+    {
+      echo self::SPANISH;
+    }
   }
 ?>
