+++ 05/24.php
@@ -1,26 +1,9 @@
 <?php
-  $object = new Son;
-  $object->test();
-  $object->test2();
-
-  class Dad
+  class User
   {
-    function test()
+    final function copyright()
     {
-      echo "[Class Dad] I am your Father<br>";
-    }
-  }
-
-  class Son extends Dad
-  {
-    function test()
-    {
-      echo "[Class Son] I am Luke<br>";
-    }
-
-    function test2()
-    {
-      parent::test();
+      echo "This class was written by Joe Smith";
     }
   }
 ?>
