--- 05/27.php
@@ -1,9 +1,0 @@
-<?php
-  class User
-  {
-    final function copyright()
-    {
-      echo "This class was written by Joe Smith";
-    }
-  }
-?>
